Add abort if unable to read lease database
Separately handle the case where the database does not exist and the
case where it may have been corrupted.  If the database does not exist
rcres will start with an empty database.  If the database cannot be
read, possibility due to corruption, rcres will abort and report the
error. Previously, all database errors were handled silently with the
database being replaced by an empty one.